+++ CHANGELOG.md
@@ -9,7 +9,8 @@
   - Display primary key tip only for array stores.
   - Fix store compilation when using multi stores registration.
   - A store cannot reference itself.
-  - Fix directive plugin on radio button
+  - Fix directive plugin on radio button.
+  - Fix ILIKE filter in offline.
 
 ## 1.10.0
 *2019-07-01*